feat: 상품 목록 조회 API 개발,
- Spring Data Jpa + Pageable 페이징 적용

+++ src/main/kotlin/scope/commerce/common/api/response/ApiResponse.kt
@@ -19,7 +19,7 @@ data class ApiResponse<T>(
     data class Pagination(
         val page: Int,
         val size: Int,
-        val totalElements: Int,
+        val totalElements: Long,
         val totalPages: Double
     )
